# Conflicts:
#	src/main/scala/com/ubirch/services/tenantadmin/TenantAdminService.scala
#	src/test/scala/com/ubirch/e2e/controllers/TenantAdminControllerSpec.scala

+++ src/main/scala/com/ubirch/controllers/TenantAdminController.scala
@@ -29,7 +29,7 @@ import monix.execution.Scheduler
 import org.joda.time.DateTime
 import org.json4s.Formats
 import org.json4s.native.Serialization
-import org.json4s.native.Serialization.{ read, write }
+import org.json4s.native.Serialization.write
 import org.scalatra._
 import org.scalatra.swagger.{ Swagger, SwaggerSupportSyntax }
 
@@ -120,6 +120,7 @@ class TenantAdminController @Inject() (
       .description("Update PoC that belong to the querying tenant.")
       .tags("Tenant-Admin", "PoC")
       .authorizations()
+
   val getPocAdmins: SwaggerSupportSyntax.OperationBuilder =
     apiOperation[String]("retrieve all poc admins of the requesting tenant")
       .summary("Get PoC admins")
@@ -229,7 +230,7 @@ class TenantAdminController @Inject() (
       getParamAsUUID("id", id => s"Invalid poc id '$id'") { id =>
         for {
           body <- readBodyWithCharset(request, StandardCharsets.UTF_8)
-          r <- tenantAdminService.updatePoc(tenant, id, read[Poc_IN](body)).map {
+          r <- tenantAdminService.updatePoc(tenant, id, Serialization.read[Poc_IN](body)).map {
             case Left(e) => e match {
                 case UpdatePocError.NotFound(pocId) =>
                   NotFound(NOK.resourceNotFoundError(s"PoC with id '$pocId' does not exist"))
@@ -238,6 +239,8 @@ class TenantAdminController @Inject() (
                     s"PoC with id '$pocId' does not belong to tenant with id '${tenantId.value.value}'"))
                 case UpdatePocError.NotCompleted(pocId, status) =>
                   Conflict(NOK.conflict(s"Poc '$pocId' is in wrong status: '$status', required: '$Completed'"))
+                case UpdatePocError.ValidationError(message) =>
+                  BadRequest(NOK.badRequest(message))
               }
             case Right(p) => Presenter.toJsonResult(p)
           }
@@ -446,38 +449,41 @@ class TenantAdminController @Inject() (
   }
 
   delete("/poc-admin/:id/2fa-token", operation(delete2FATokenOnPocAdmin)) {
-    tenantAdminEndpoint("Delete 2FA token for PoC admin") { tenant =>
-      getParamAsUUID("id", id => s"Invalid PocAdmin id '$id'") { pocAdminId =>
-        for {
-          maybePocAdmin <- pocAdminRepository.getPocAdmin(pocAdminId)
-          notFoundMessage = s"Poc admin with id '$pocAdminId' not found'"
-          r <- maybePocAdmin match {
-            case None => Task.pure(NotFound(NOK.resourceNotFoundError(notFoundMessage)))
-            case Some(pocAdmin) if pocAdmin.tenantId != tenant.id =>
-              Task.pure(NotFound(NOK.resourceNotFoundError(notFoundMessage)))
-            case Some(pocAdmin) if pocAdmin.status != Completed =>
-              Task.pure(Conflict(NOK.conflict(
-                s"Poc admin '$pocAdminId' is in wrong status: '${pocAdmin.status}', required: '${Completed}'")))
-            case Some(pocAdmin) => certifyUserService.remove2FAToken(CertifyKeycloak.defaultRealm, pocAdmin)
-                .flatMap {
-                  case Left(e) => e match {
-                      case Remove2faTokenError.KeycloakError(_, error) =>
-                        error match {
-                          case Remove2faTokenKeycloakError.UserNotFound(error) =>
-                            Task.pure(NotFound(NOK.resourceNotFoundError(error)))
-                          case Remove2faTokenKeycloakError.KeycloakError(error) =>
-                            Task.pure(InternalServerError(NOK.serverError(error)))
-                        }
-                      case Remove2faTokenError.MissingCertifyUserId(id) =>
-                        Task.pure(Conflict(NOK.conflict(s"Poc admin '$id' does not have certifyUserId")))
-                    }
-                  case Right(_) =>
-                    pocAdminRepository.updatePocAdmin(pocAdmin.copy(webAuthnDisconnected =
-                      Some(DateTime.parse(clock.instant().toString)))) >>
-                      Task.pure(Ok(""))
-                }
-          }
-        } yield r
+    if (true) NotFound(NOK.noRouteFound("Sorry, this method is not yet fully implemented."))
+    else {
+      tenantAdminEndpoint("Delete 2FA token for PoC admin") { tenant =>
+        getParamAsUUID("id", id => s"Invalid PocAdmin id '$id'") { pocAdminId =>
+          for {
+            maybePocAdmin <- pocAdminRepository.getPocAdmin(pocAdminId)
+            notFoundMessage = s"Poc admin with id '$pocAdminId' not found'"
+            r <- maybePocAdmin match {
+              case None => Task.pure(NotFound(NOK.resourceNotFoundError(notFoundMessage)))
+              case Some(pocAdmin) if pocAdmin.tenantId != tenant.id =>
+                Task.pure(NotFound(NOK.resourceNotFoundError(notFoundMessage)))
+              case Some(pocAdmin) if pocAdmin.status != Completed =>
+                Task.pure(Conflict(NOK.conflict(
+                  s"Poc admin '$pocAdminId' is in wrong status: '${pocAdmin.status}', required: '${Completed}'")))
+              case Some(pocAdmin) => certifyUserService.remove2FAToken(CertifyKeycloak.defaultRealm, pocAdmin)
+                  .flatMap {
+                    case Left(e) => e match {
+                        case Remove2faTokenError.KeycloakError(_, error) =>
+                          error match {
+                            case Remove2faTokenKeycloakError.UserNotFound(error) =>
+                              Task.pure(NotFound(NOK.resourceNotFoundError(error)))
+                            case Remove2faTokenKeycloakError.KeycloakError(error) =>
+                              Task.pure(InternalServerError(NOK.serverError(error)))
+                          }
+                        case Remove2faTokenError.MissingCertifyUserId(id) =>
+                          Task.pure(Conflict(NOK.conflict(s"Poc admin '$id' does not have certifyUserId")))
+                      }
+                    case Right(_) =>
+                      pocAdminRepository.updatePocAdmin(pocAdmin.copy(webAuthnDisconnected =
+                        Some(DateTime.parse(clock.instant().toString)))) >>
+                        Task.pure(Ok(""))
+                  }
+            }
+          } yield r
+        }
       }
     }
   }
